+++ addOns/grpc/CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Update dependency.
 - Maintenance changes.
 
+### Fixed
+- Handle native binary gRPC messages in the message view and active scanner.
+
 ## [0.2.0] - 2024-07-02
 
 ### Added